requirements.txt
# Core dependencies
discord.py
httpx
openai
pyyaml
pytz
python-dotenv

# LangGraph core dependencies
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Google AI and search APIs
google-generativeai>=0.8.0
google-api-python-client>=2.0.0

# Data processing and validation
pydantic>=2.0.0
typing-extensions>=4.0.0

# Async support
asyncio
aiofiles

# Scheduling
apscheduler>=3.10.0

# Development and debugging tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
# jupyter>=1.0.0